requirements.txt
Flask==2.3.2
gunicorn==20.1.0
requests==2.31.0
pyairtable==2.3.3
python-dotenv==1.0.0